move control character escape upstream
